refactor(consent): split the transmit decision's persistence out of authorizetransmit.go
The file crossed the 500-line ceiling once deliveryLinks was threaded
through. recordDecisions, the suppression reads and decisionRecipientKey
move to authorizetransmitrecord.go, the same split authorizedecide.go
already made from this file for the same reason.
